--- kafka-connect-s3/src/main/java/io/confluent/connect/s3/format/orc/schema/OrcLongField.java
@@ -1,22 +1,0 @@
-package io.confluent.connect.s3.format.orc.schema;
-
-import org.apache.hadoop.hive.ql.exec.vector.ColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.LongColumnVector;
-
-class OrcLongField extends AbstractOrcField {
-
-  OrcLongField(String name, int columnIndex, boolean optional) {
-    super(name, columnIndex, optional);
-  }
-
-  @Override
-  protected void setNotNullData(ColumnVector[] orcColumns, Object data, int rowIndex) {
-    LongColumnVector col = (LongColumnVector) orcColumns[columnIndex];
-    col.vector[rowIndex] = (Long) data;
-  }
-
-  @Override
-  protected String getTypeDescription() {
-    return "bigint";
-  }
-}
